+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Uncommited
+
+### Added 
+- `get_val_mask_func_kwargs` and  `get_trainer_func_kwargs` kwargs for corresponding functions 
+
+
 ## [0.13.0] - 10.04.2025
 
 ### Added